CONN-10484 Obtaining RowSchema from Rowset API
CONN-10463 Interface

CONN-10484 Repackaging - avoid using internal sdk classes in the connector
