+++ src/main/java/com/snowflake/kafka/connector/internal/streaming/validation/RowSchema.java
@@ -2,6 +2,7 @@ package com.snowflake.kafka.connector.internal.streaming.validation;
 
 import static net.snowflake.ingest.utils.Utils.concatDotPath;
 
+import com.snowflake.kafka.connector.internal.streaming.common.ColumnProperties;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
