+++ src/main/java/com/snowflake/kafka/connector/internal/streaming/validation/RowsetApiRowSchemaProvider.java
@@ -66,7 +66,6 @@ public class RowsetApiRowSchemaProvider implements RowSchemaProvider {
     metadata.setCollation(col.getCollation());
     metadata.setType(dataType.getType());
     metadata.setLogicalType(dataType.getType());
-    metadata.setSourceIcebergDataType(col.getSourceIcebergDataType());
     if (dataType.getByteLength() != null) {
       metadata.setByteLength(dataType.getByteLength());
     }
